added admin role and check on delete api call

+++ src/server/api/routers/item.ts
@@ -1,21 +1,26 @@
-import { date, z } from "zod";
+import { z } from "zod";
 
 import { createTRPCRouter, publicProcedure, protectedProcedure } from "~/server/api/trpc";
 import { ItemSchema } from "~/components/ItemCRUD";
 
 export const itemRouter = createTRPCRouter({
-  // getMine: protectedProcedure
-  //   .input(z.number())
-  //   .query(({ ctx }) => {
-  //     return ctx.prisma.item.findMany({
-  //       where: {
-  //       }
-  //     });
-  //   }),
+  getMine: protectedProcedure
+    .query(({ ctx }) => {
+      return ctx.prisma.item.findMany({
+        where: {
+          userId: ctx.session.user.id
+        }
+      });
+    }),
 
   getAll: publicProcedure
     .query(async ({ ctx }) => {
-       return await ctx.prisma.item.findMany();
+      return await ctx.prisma.item.findMany({
+        include: {
+          user: true
+        },
+      }
+      );
     }),
 
   insertOne: protectedProcedure
@@ -25,14 +30,19 @@ export const itemRouter = createTRPCRouter({
         data: {
           name: input.name,
           property: input.property,
-          date: input.date
+          date: input.date,
+          userId: ctx.session.user.id
         }
       });
     }),
 
   deleteOne: protectedProcedure
     .input(z.number())
-    .mutation(({ ctx, input }) => {
+    .mutation( async ({ ctx, input }) => {
+      const itemToDelete = await ctx.prisma.item.findUnique({where: { id: input }});
+      if (itemToDelete?.userId !== ctx.session.user.id && ctx.session.user.role !== "ADMIN") {
+        throw new Error("You are not authorized to delete this item");
+      }
       return ctx.prisma.item.delete({
         where: {
           id: input,
